Make things compile under Win32 once more. Fix some warnings.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@78929 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/standard/mod_cgi.c
@@ -458,7 +458,7 @@ static int cgi_handler (request_rec *r)
 	    }
 	    reset_timeout(r);
 	    if (bwrite(script_out, argsbuffer, len_read)
-	            < (size_t)len_read) {
+	            < len_read) {
 	        /* silly script stopped reading, soak up remaining message */
 	        while (get_client_block(r, argsbuffer, HUGE_STRING_LEN) > 0)
 	            ; /* dump it */
